Audio: MFCC: Split the implementation of mfcc into 2 files
Split the implementation code of mfcc from mfcc_generic.c
into mfcc_generic,c and mfcc_common.c.
