+++ src/audio/mfcc/CMakeLists.txt
@@ -1,3 +1,3 @@
 # SPDX-License-Identifier: BSD-3-Clause
 
-add_local_sources(sof mfcc.c mfcc_setup.c mfcc_generic.c)
+add_local_sources(sof mfcc.c mfcc_setup.c mfcc_common.c mfcc_generic.c)
